+++ askbot/views/writers.py
@@ -35,12 +35,14 @@ from django.views.decorators import csrf
 from askbot import exceptions as askbot_exceptions
 from askbot import forms
 from askbot import models
+from askbot.models import signals
 from askbot.conf import settings as askbot_settings
 from askbot.utils import decorators
 from askbot.utils.forms import format_errors
 from askbot.utils.functions import diff_date
 from askbot.utils import url_utils
 from askbot.utils.file_utils import store_file
+from askbot.utils.loading import load_module
 from askbot.views import context
 from askbot.templatetags import extra_filters_jinja as template_filters
 from askbot.importers.stackexchange import management as stackexchange#todo: may change
@@ -60,7 +62,6 @@ ANSWERS_PAGE_SIZE = 10
 def upload(request):#ajax upload file to a question or answer
     """view that handles file upload via Ajax
     """
-
     # check upload permission
     result = ''
     error = ''
@@ -79,10 +80,11 @@ def upload(request):#ajax upload file to a question or answer
             raise exceptions.PermissionDenied('invalid upload file name prefix')
 
         #todo: check file type
-        f = request.FILES['file-upload']#take first file
+        uploaded_file = request.FILES['file-upload']#take first file
+        orig_file_name = uploaded_file.name
         #todo: extension checking should be replaced with mimetype checking
         #and this must be part of the form validation
-        file_extension = os.path.splitext(f.name)[1].lower()
+        file_extension = os.path.splitext(orig_file_name)[1].lower()
         if not file_extension in settings.ASKBOT_ALLOWED_UPLOAD_FILE_TYPES:
             file_types = "', '".join(settings.ASKBOT_ALLOWED_UPLOAD_FILE_TYPES)
             msg = _("allowed file types are '%(file_types)s'") % \
@@ -91,7 +93,7 @@ def upload(request):#ajax upload file to a question or answer
 
         # generate new file name and storage object
         file_storage, new_file_name, file_url = store_file(
-                                            f, file_name_prefix
+                                            uploaded_file, file_name_prefix
                                         )
         # check file size
         # byte
@@ -120,8 +122,8 @@ def upload(request):#ajax upload file to a question or answer
     #    'file_url': file_url
     #})
     #return HttpResponse(data, mimetype = 'application/json')
-    xml_template = "<result><msg><![CDATA[%s]]></msg><error><![CDATA[%s]]></error><file_url>%s</file_url></result>"
-    xml = xml_template % (result, error, file_url)
+    xml_template = "<result><msg><![CDATA[%s]]></msg><error><![CDATA[%s]]></error><file_url>%s</file_url><orig_file_name><![CDATA[%s]]></orig_file_name></result>"
+    xml = xml_template % (result, error, file_url, orig_file_name)
 
     return HttpResponse(xml, mimetype="application/xml")
 
@@ -479,6 +481,13 @@ def edit_question(request, id):
 def edit_answer(request, id):
     answer = get_object_or_404(models.Post, id=id)
     revision = answer.get_latest_revision()
+
+    class_path = getattr(settings, 'ASKBOT_EDIT_ANSWER_FORM', None)
+    if class_path:
+        edit_answer_form_class = load_module(class_path)
+    else:
+        edit_answer_form_class = forms.EditAnswerForm
+
     try:
         request.user.assert_can_edit_answer(answer)
         if request.method == "POST":
@@ -493,18 +502,18 @@ def edit_answer(request, id):
                     # Replace with those from the selected revision
                     rev = revision_form.cleaned_data['revision']
                     revision = answer.revisions.get(revision = rev)
-                    form = forms.EditAnswerForm(
+                    form = edit_answer_form_class(
                                     answer, revision, user=request.user
                                 )
                 else:
-                    form = forms.EditAnswerForm(
-                                            answer,
-                                            revision,
-                                            request.POST,
-                                            user=request.user
-                                        )
+                    form = edit_answer_form_class(
+                                                answer,
+                                                revision,
+                                                request.POST,
+                                                user=request.user
+                                            )
             else:
-                form = forms.EditAnswerForm(
+                form = edit_answer_form_class(
                     answer, revision, request.POST, user=request.user
                 )
                 revision_form = forms.RevisionForm(answer, revision)
@@ -522,12 +531,27 @@ def edit_answer(request, id):
                             is_private=is_private,
                             suppress_email=suppress_email
                         )
+
+                        signals.answer_edited.send(None,
+                            answer=answer,
+                            user=user,
+                            form_data=form.cleaned_data
+                        )
+
                     return HttpResponseRedirect(answer.get_absolute_url())
         else:
             revision_form = forms.RevisionForm(answer, revision)
-            form = forms.EditAnswerForm(answer, revision, user=request.user)
+            form = edit_answer_form_class(answer, revision, user=request.user)
             if request.user.can_make_group_private_posts():
                 form.initial['post_privately'] = answer.is_private()
+
+        #gives a chance to set extra initial data on the form
+        signals.answer_before_editing.send(None,
+            answer=answer,
+            user=request.user,
+            form=form
+        )
+
         data = {
             'page_class': 'edit-answer-page',
             'active_tab': 'questions',
@@ -555,7 +579,15 @@ def answer(request, id):#process a new answer
     """
     question = get_object_or_404(models.Post, post_type='question', id=id)
     if request.method == "POST":
-        form = forms.AnswerForm(request.POST, user=request.user)
+
+        custom_class_path = getattr(settings, 'ASKBOT_NEW_ANSWER_FORM', None)
+        if custom_class_path:
+            form_class = load_module(custom_class_path)
+        else:
+            form_class = forms.AnswerForm
+
+        form = form_class(request.POST, user=request.user)
+
         if form.is_valid():
             wiki = form.cleaned_data['wiki']
             text = form.cleaned_data['text']
@@ -581,6 +613,13 @@ def answer(request, id):#process a new answer
                                         is_private = is_private,
                                         timestamp = update_time,
                                     )
+
+                    signals.new_answer_posted.send(None,
+                        answer=answer,
+                        user=user,
+                        form_data=form.cleaned_data
+                    )
+
                     return HttpResponseRedirect(answer.get_absolute_url())
                 except askbot_exceptions.AnswerAlreadyGiven, e:
                     request.user.message_set.create(message = unicode(e))
@@ -619,7 +658,7 @@ def __generate_comments_json(obj, user):#non-view generates json data for the po
                 is_deletable = True
             except exceptions.PermissionDenied:
                 is_deletable = False
-            is_editable = template_filters.can_edit_comment(comment.author, comment)
+            is_editable = template_filters.can_edit_comment(user, comment)
         else:
             is_deletable = False
             is_editable = False
@@ -648,6 +687,10 @@ def __generate_comments_json(obj, user):#non-view generates json data for the po
 @csrf.csrf_exempt
 @decorators.check_spam('comment')
 def post_comments(request):#generic ajax handler to load comments to an object
+    """todo: fixme: post_comments is ambigous:
+    means either get comments for post or 
+    add a new comment to post
+    """
     # only support get post comments by ajax now
 
     post_type = request.REQUEST.get('post_type', '')
@@ -656,11 +699,27 @@ def post_comments(request):#generic ajax handler to load comments to an object
 
     user = request.user
 
-    id = request.REQUEST['post_id']
-    obj = get_object_or_404(models.Post, id=id)
+    if request.method == 'POST':
+        form = forms.NewCommentForm(request.POST)
+    elif request.method == 'GET':
+        form = forms.GetCommentsForPostForm(request.GET)
+
+    if form.is_valid() == False:
+        return HttpResponseBadRequest(
+            _('This content is forbidden'),
+            mimetype='application/json'
+        )
+
+    post_id = form.cleaned_data['post_id']
+    try:
+        post = models.Post.objects.get(id=post_id)
+    except models.Post.DoesNotExist:
+        return HttpResponseBadRequest(
+            _('Post not found'), mimetype='application/json'
+        )
 
     if request.method == "GET":
-        response = __generate_comments_json(obj, user)
+        response = __generate_comments_json(post, user)
     elif request.method == "POST":
         try:
             if user.is_anonymous():
@@ -669,47 +728,54 @@ def post_comments(request):#generic ajax handler to load comments to an object
                         '<a href="%(sign_in_url)s">sign in</a>.') % \
                         {'sign_in_url': url_utils.get_login_url()}
                 raise exceptions.PermissionDenied(msg)
-            user.post_comment(parent_post=obj, body_text=request.POST.get('comment'))
-            response = __generate_comments_json(obj, user)
+            user.post_comment(
+                parent_post=post, body_text=form.cleaned_data['comment']
+            )
+            response = __generate_comments_json(post, user)
         except exceptions.PermissionDenied, e:
             response = HttpResponseForbidden(unicode(e), mimetype="application/json")
 
     return response
 
-@csrf.csrf_exempt
+#@csrf.csrf_exempt
 @decorators.ajax_only
-@decorators.check_spam('comment')
+#@decorators.check_spam('comment')
 def edit_comment(request):
     if request.user.is_anonymous():
         raise exceptions.PermissionDenied(_('Sorry, anonymous users cannot edit comments'))
 
     form = forms.EditCommentForm(request.POST)
     if form.is_valid() == False:
-        return HttpResponseBadRequest()
-        
-    comment_id = form.cleaned_data['comment_id']
-    suppress_email = form.cleaned_data['suppress_email']
+        raise exceptions.PermissionDenied('This content is forbidden')
 
-    comment_post = models.Post.objects.get(post_type='comment', id=comment_id)
+    comment_post = models.Post.objects.get(
+                    post_type='comment',
+                    id=form.cleaned_data['comment_id']
+                )
 
     request.user.edit_comment(
         comment_post=comment_post,
-        body_text = request.POST['comment'],
-        suppress_email=suppress_email
+        body_text=form.cleaned_data['comment'],
+        suppress_email=form.cleaned_data['suppress_email']
     )
 
-    is_deletable = template_filters.can_delete_comment(comment_post.author, comment_post)
-    is_editable = template_filters.can_edit_comment(comment_post.author, comment_post)
+    is_deletable = template_filters.can_delete_comment(
+                            comment_post.author, comment_post)
+
+    is_editable = template_filters.can_edit_comment(
+                            comment_post.author, comment_post)
+
     tz = ' ' + template_filters.TIMEZONE_STR
 
     tz = template_filters.TIMEZONE_STR
+    timestamp = str(comment_post.added_at.replace(microsecond=0)) + tz
 
     return {
         'id' : comment_post.id,
         'object_id': comment_post.parent.id,
-        'comment_added_at': str(comment_post.added_at.replace(microsecond = 0)) + tz,
+        'comment_added_at': timestamp,
         'html': comment_post.html,
-        'user_display_name': comment_post.author.username,
+        'user_display_name': escape(comment_post.author.username),
         'user_url': comment_post.author.get_profile_url(),
         'user_id': comment_post.author.id,
         'is_deletable': is_deletable,